requirements.txt
# Core ML/DL
torch>=2.1.0
torchvision>=0.16.0
transformers[torch]>=4.35.0
accelerate>=0.25.0
sentencepiece>=0.1.99

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.1

# OCR & Document AI
pytesseract>=0.3.10
easyocr>=1.7.0
pdf2image>=1.16.3

# Vision-Language Models (Qwen-VL)
qwen-vl-utils>=0.0.8
einops>=0.7.0
timm>=0.9.12

# Agent Framework
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# API & Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration & Logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Monitoring & Metrics
prometheus-client>=0.19.0
tensorboard>=2.15.0

# Semi-Supervised Learning
scipy>=1.11.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
huggingface-hub>=0.19.0

# Google Generative AI
langchain-google-genai>=2.0.0
google-generativeai>=0.3.0